form without minimal style, and no validation
