+++ client/package.json
@@ -33,7 +33,8 @@
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.1",
     "titleize": "^2.1.0",
-    "uppercamelcase": "^3.0.0"
+    "uppercamelcase": "^3.0.0",
+    "yup": "^0.29.0"
   },
   "scripts": {
     "start": "react-scripts start",
