+++ client/src/components/form/Form.css
@@ -54,15 +54,6 @@
   padding: 15px;
 }
 
-.form-wrapper .form .containers .section .container .member {
-  background-color: #eceff1;
-  border: #b6b6b6 1px solid;
-  border-radius: 30px;
-  padding: 4px 11px;
-  margin: 4px 5px;
-  cursor: pointer;
-}
-
 .form-wrapper .form .containers .section .container .member:hover {
   background-color: #b6b6b6;
 }
@@ -90,35 +81,33 @@
 .form-wrapper .form .input-area {
   -ms-flex-preferred-size: 50%;
       flex-basis: 50%;
-  text-align: center;
   background-color: #e8eaf6;
-}
-
-.form-wrapper .form .input-area .inputs {
   display: -ms-grid;
   display: grid;
-  -ms-grid-columns: (1fr)[2];
-      grid-template-columns: repeat(2, 1fr);
+  -ms-grid-columns: (30px)[auto];
+      grid-template-columns: repeat(auto, 30px);
   -ms-grid-rows: (1fr)[2];
       grid-template-rows: repeat(2, 1fr);
 }
 
-.form-wrapper .form .input-area .inputs .input-field {
+.form-wrapper .form .input-area .input-field {
   margin: 5px;
 }
 
-.form-wrapper .form .input-area .inputs .input-field .input-label {
+.form-wrapper .form .input-area .input-field label {
   text-align: left;
   padding-left: 10px;
+  width: 100px;
 }
 
-.form-wrapper .form .input-area .inputs .block {
-  grid-row: span 2;
+.form-wrapper .form .input-area .input-field input {
+  width: 80%;
 }
 
-.form-wrapper .form .input-area .inputs input {
+.form-wrapper .form .input-area input,
+.form-wrapper .form .input-area select {
   border: #b6b6b6 1px solid;
-  border-radius: 30px;
+  border-radius: 5px;
   padding: 4px 11px;
   margin: 4px 5px;
   width: 90%;
@@ -126,7 +115,7 @@
   font-family: "Segoe UI";
 }
 
-.form-wrapper .form .input-area .inputs textarea {
+.form-wrapper .form .input-area textarea {
   border: #b6b6b6 1px solid;
   border-radius: 5px;
   padding: 4px 11px;
@@ -137,7 +126,20 @@
   font-family: "Segoe UI";
 }
 
-.form-wrapper .form .input-area .header {
-  font-size: 24px;
+.form-wrapper .form .input-area #submit {
+  font-size: 16px;
+}
+
+.member {
+  background-color: #eceff1;
+  border: #b6b6b6 1px solid;
+  border-radius: 30px;
+  padding: 4px 11px;
+  margin: 4px 5px;
+  cursor: pointer;
+}
+
+.member:hover {
+  background-color: #b6b6b6;
 }
 /*# sourceMappingURL=Form.css.map */
