+++ client/src/components/form/InputArea.js
@@ -1,16 +1,14 @@
-import React, { Component } from "react";
-import InputField from "./InputField";
-import { capitalCase, paramCase } from "change-case";
+import React from "react";
+import { Formik, Form, Field, FieldArray } from "formik";
+import { capitalCase } from "change-case";
+import { paramCase } from "change-case";
 const axios = require("axios");
-var _ = require("lodash");
 
-export class InputArea extends Component {
+class InputArea extends React.Component {
   constructor(props) {
     super(props);
-
     this.state = {
       riverList: null,
-      dataArr: this.props.dataArr,
     };
   }
 
@@ -32,54 +30,97 @@ export class InputArea extends Component {
     }
   }
 
-  handleChange = (e) => {
-    
-  };
-
   render() {
-    const { river, rapid, feature, featureName } = this.props;
-    
-    let dataArr = this.state.dataArr ? [...this.state.dataArr] : null
-
-    const data =
-      feature !== null
-        ? dataArr[river].rapids[rapid][featureName][feature]
-        : rapid !== null
-        ? dataArr[river].rapids[rapid]
-        : river !== null
-        ? dataArr[river]
-        : null;
-
-    let inputs = [];
-    let key = 0;
-
-    const input = (data, prevProp) => {
-      for (let elem in data) {
-        key++;
-        if (typeof data[elem] == "object" && !Array.isArray(data[elem])) {
-          input(data[elem], elem);
-        } else if (!Array.isArray(data[elem]) || elem == "range") {
-          inputs.push(
-            <InputField
-              prop={
-                prevProp.length == 0
-                  ? `${capitalCase(elem)}`
-                  : `${capitalCase(prevProp)} - ${capitalCase(elem)}`
-              }
-              value={data[elem]}
-              key={`field${key}`}
-              handleChange={this.handleChange}
-            />
-          );
-        }
-      }
-    };
-
-    input(data, "");
+    const { river, rapid, feature, featureName, dataArr } = this.props;
 
     return (
       <div className="input-area">
-        <div className="inputs">{inputs}</div>
+        <Formik
+          initialValues={dataArr} //set values:_________
+          onSubmit={(values) => {
+            console.log(values);
+          }}
+        >
+          {({ values, handleChange, handleSubmit, isSubmitting }) => {
+            const data =
+              feature !== null
+                ? values[river].rapids[rapid][featureName][feature]
+                : rapid !== null
+                ? values[river].rapids[rapid]
+                : river !== null
+                ? values[river]
+                : null;
+
+            let name =
+              feature !== null
+                ? `[${river}][rapids][${rapid}][${featureName}][${feature}]`
+                : rapid !== null
+                ? `[${river}][rapids][${rapid}]`
+                : river !== null
+                ? `[${river}]`
+                : null;
+            let list = [];
+            const parseObject = (dataObj, name) => {
+              for (let elem in dataObj) {
+                name = `${name}.${elem}`;
+                if (
+                  typeof dataObj[elem] == "object" &&
+                  !Array.isArray(dataObj[elem])
+                ) {
+                  parseObject(dataObj[elem], name);
+                } else if (!Array.isArray(data[elem]) || elem == "range") {
+                  if (elem == "pointerDirection") {
+                    list.push(
+                      <div className="input-field">
+                        <div>{capitalCase(elem)}:</div>
+                        <Field as="select" name={name} key={elem}>
+                          <option value="top">Top</option>
+                          <option value="bottom">Bottom</option>
+                          <option value="left">Left</option>
+                          <option value="right">Right</option>
+                        </Field>
+                      </div>
+                    );
+                  } else if (elem == "path") {
+                    let options = this.state.riverList.map((path) => {
+                      path = path.split(".")[0];
+                      return <option value={path}>{path}</option>;
+                    });
+                    list.push(
+                      <div className="input-field">
+                        <div>{capitalCase(elem)}:</div>
+                        <Field as="select" name={name} key={elem}>
+                          {options}
+                        </Field>
+                      </div>
+                    );
+                  } else {
+                    list.push(
+                      <div className="input-field">
+                        <div>{capitalCase(elem)}:</div>
+                        <Field name={name} key={elem} />
+                      </div>
+                    );
+                  }
+                }
+
+                let temp = name.split(".");
+                temp.pop();
+                name = temp.join(".");
+              }
+              return list;
+            };
+
+            return (
+              <Form>
+                {parseObject(data, name)}
+                <button type="submit" className="member" id="submit">
+                  Submit
+                </button>
+              </Form>
+            );
+          }}
+        </Formik>
       </div>
     );
   }
